feature - course_size report update.

+++ report/coursesize/classes/task/report_async.php
@@ -46,27 +46,44 @@ class report_async extends \core\task\scheduled_task {
         global $DB, $CFG;
         require_once($CFG->dirroot . '/report/coursesize/locallib.php');
 
+        // BEGIN LSU - Store course size and history.
+        // Are we using cron or no?
         if (get_config('report_coursesize', 'calcmethod') == 'cron') {
 
-            mtrace("Generating report_coursesize cache...");
+            // If we want to store historical data then don't purge then insert.
+            // Insert with the timestamp as of NOW.
             set_time_limit(0);
-
-            // First we delete the old data, then we re-populate it, wrap in a transaction to help keep it together.
+            
+            $processtime = time();
             $transaction = $DB->start_delegated_transaction();
+            $historysql = " AND rc.timestamp = ".$processtime;
+            // If we want to store historical data then don't purge then insert.
+            // Insert with the timestamp as of NOW.
+            if ((int)get_config('report_coursesize', 'keephistory') != 1) {
+                // First we delete the old data, then we re-populate it,
+                // wrap in a transaction to help keep it together.
+                $DB->delete_records('report_coursesize');
+            }
 
-            // Clean up cache table.
-            $DB->delete_records('report_coursesize');
-
+            mtrace("Generating report_coursesize data...");
+            
             // Generate report_coursesize table.
-            $basesql = report_coursesize_filesize_sql();
-            $sql = "INSERT INTO {report_coursesize} (course, filesize) $basesql ";
-            $DB->execute($sql);
+            $basesql = report_coursesize_filesize_sql($processtime);
+
+            $sql = "INSERT INTO {report_coursesize} (course, filesize, timestamp) $basesql ";
+            $DB->execute($sql, array($processtime));
 
             // Now calculate size of backups.
             $basesql = report_coursesize_backupsize_sql();
 
             $sql = "UPDATE {report_coursesize} rc
-                    SET backupsize = (SELECT bf.filesize FROM ($basesql) bf WHERE bf.course = rc.course)";
+                SET backupsize = (
+                    SELECT bf.filesize FROM ($basesql) bf
+                    WHERE bf.course = rc.course
+                    $historysql
+            )";
+
+            // END LSU - Store course size and history.
             $DB->execute($sql);
 
             $transaction->allow_commit();
@@ -76,7 +93,9 @@ class report_async extends \core\task\scheduled_task {
             set_config('coursesizeupdated', time(), 'report_coursesize');
 
             mtrace("report_coursesize cache updated.");
+            
         }
+
         // Check if the path ends with a "/" otherwise an exception will be thrown.
         $sitedatadir = $CFG->dataroot;
         if (is_dir($sitedatadir)) {
@@ -87,7 +106,12 @@ class report_async extends \core\task\scheduled_task {
         }
 
         // Total files usage either hasn't been stored, or is out of date.
+        mtrace("Getting moodledata size.");
+        $tustime = microtime(true);
         $totalusage = get_directory_size($sitedatadir);
+        $tuetime = microtime(true);
+        $tuelapsed = round($tuetime - $tustime, 1);
+        mtrace("Retreived moodledata size of $totalusage bytes in $tuelapsed seconds.");
         set_config('filessize', $totalusage, 'report_coursesize');
         set_config('filessizeupdated', time(), 'report_coursesize');
